+++ scripts/devops_tasks/test_run_samples.py
@@ -144,8 +144,10 @@ IGNORED_SAMPLES = {
     "azure-ai-textanalytics": [
         "sample_analyze_healthcare_entities_with_cancellation.py",
         "sample_analyze_healthcare_entities_with_cancellation_async.py",
-        "sample_abstractive_summary.py",
-        "sample_abstractive_summary_async.py",
+        "sample_abstract_summary.py",
+        "sample_abstract_summary_async.py",
+        "sample_authentication.py",
+        "sample_authentication_async.py",
     ],
     "azure-ai-language-conversations": [
         "sample_assign_deployment_resources.py",