+++ sdk/ml/azure-ai-ml/azure/ai/ml/_internal/_setup.py
@@ -58,7 +58,6 @@ def _register_node(_type, node_cls, schema_cls):
 
 
 def enable_internal_components_in_pipeline(*, force=False):
-    global _registered  # pylint: disable=global-statement
     if _registered and not force:
         return  # already registered